setting Context
