+++ client/App.tsx
@@ -1,10 +1,39 @@
 import React, { useState, useEffect, useContext } from "react";
 import SearchTable from "./SearchTable";
-import Dropdowns from "./Dropdowns";
+import Dropdowns from "./dropdowns";
 import CrimeMap from "./CrimeMap";
 import { CrimesContextProvider } from "./CrimesContext"; // lint error says this implicitly has an "any" type
 import Authentication from "./Authentication/Authentication";
 import Footer from "./Footer";
+
+export interface Crime {
+  id: string,
+  case_number: string,
+  date: string,
+  block: string,
+  iucr: string,
+  primary_type: string,
+  description: string,
+  location_description: string,
+  arrest: boolean,
+  domestic: boolean,
+  beat: string,
+  district: string,
+  ward: string,
+  community_area: string,
+  fbi_code: string,
+  x_coordinate: string,
+  y_coordinate: string,
+  year: string,
+  updated_on: string,
+  latitude: string,
+  longitude: string,
+  location: {
+      latitude: string,
+      longitude: string,
+      human_address: string
+  }
+}
 
 const App = () => {
   return (
