+++ client/CrimesContext.js
@@ -337,7 +337,7 @@ const crimesMockData = [
     }
 ]
 
-const CrimesContext = createContext(); //setting CrimesContext as the context
+const CrimesContext = createContext(crimesMockData); //setting CrimesContext as the context
 
 const CrimesContextProvider = (props) => { //this acts as the wrapper for the components who must use the same state (essentially as a customized global scope, since all the children and children's children should be able to access this correctly without drilling- fingers crossed)
