+++ package.json
@@ -32,7 +32,6 @@
     "cors": "^2.8.5",
     "express": "^4.17.3",
     "firebase": "^9.6.10",
-    "jquery": "^3.6.0",
     "leaflet": "^1.7.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
@@ -44,7 +43,6 @@
     "@babel/core": "^7.17.8",
     "@babel/preset-react": "^7.16.7",
     "@types/cors": "^2.8.12",
-    "@types/jquery": "^3.5.14",
     "@types/leaflet": "^1.7.9",
     "@types/node": "^17.0.23",
     "@types/react": "^17.0.43",
