+++ server/index.ts
@@ -4,17 +4,19 @@ import cors from 'cors';
 const port = 8081;
 const app: Express = express();
 
-import { Crime } from '../database/queries'
+// import { Crime } from '../client/App'
 
 app.use(express.json());
 app.use(cors());
 
 
 app.get('/crimes', (Request, Response) => {
-  console.log('REQUEST IN SERVER', Request);
+  // console.log('REQUEST IN SERVER', Request);
+
   // .then((res) => {
   // console.log('RESPONSE IN SERVER', res);
-  //   Response.send(res);
+  console.log('response in server??', Response)
+    Response.send();
   // })
 });
